+++ src/screens/WalletScreen/Content/Content.tsx
@@ -1,11 +1,8 @@
 import React, { useCallback, useEffect, useState } from 'react';
 import { TabLayout } from '@layouts';
-import { useNavigation, useTransactions, useLanguage } from '@hooks';
+import { useNavigation, useTransactions, useLanguage, useAccountName } from '@hooks';
 import { PendingTransaction } from '@components';
-import { getProvider, ZapperTransaction, getENSAddress, smallWalletAddress } from '@src/model/wallet';
-import crypto from 'crypto';
-import { INTERCOM_KEY } from '@env';
-import Intercom from '@intercom/intercom-react-native';
+import { getProvider, ZapperTransaction } from '@src/model/wallet';
 import { AssetsPanel, ActionsPanel, Header } from '../components';
 import { Transactions, Accounts } from '../screens';
 import { ContentProps } from './Content.types';
@@ -23,11 +20,11 @@ export const Content: React.FC<ContentProps> = ({
 	setAddFundsVisible,
 	setSendModalOpen
 }) => {
-	const [ensName, setEnsName] = React.useState<string | null>('');
 	const { i18n } = useLanguage();
 	const navigation = useNavigation();
 	const { loading, fetchTransactions, pendingTransactions, updatePendingTransaction } = useTransactions();
 	const [tx, setTx] = useState<ZapperTransaction | null>();
+	const accountName = useAccountName(address);
 
 	useEffect(() => {
 		const fetchStatus = async () => {
@@ -45,35 +42,10 @@ export const Content: React.FC<ContentProps> = ({
 		fetchStatus();
 	}, [pendingTransactions]);
 
-	useEffect(() => {
-		const fetchENSAddress = async () => {
-			const name = await getENSAddress(address);
-			setEnsName(name);
-		};
-
-		fetchENSAddress();
-	}, []);
-
-	const accountName = () => {
-		if (ensName) {
-			return ensName;
-		}
-		return smallWalletAddress(address);
-	};
-
 	const handleRefresh = useCallback(() => {
 		fetchTransactions();
 		setTx(null);
 	}, [fetchTransactions]);
-
-	// Register user on Intercom
-	const intercomKey = INTERCOM_KEY || process.env.INTERCOM_KEY;
-	const hmac = crypto.createHmac('sha256', intercomKey!);
-	hmac.update(accountName());
-	const sign = hmac.digest('hex');
-	Intercom.setUserHash(sign);
-	Intercom.registerIdentifiedUser({ userId: accountName() });
-	//
 
 	return (
 		<TabLayout
@@ -87,7 +59,7 @@ export const Content: React.FC<ContentProps> = ({
 			<Header
 				onSettingsPress={onSettingsPress}
 				onCopyPress={onCopyToClipboard}
-				accountName={accountName()}
+				accountName={accountName}
 			/>
 
 			{!!tx && <PendingTransaction transaction={tx} />}
